perf: make conditionalHandler more clear
